requirements.txt
# ============================================================================
# PropVision-AI Requirements
# Python 3.11+ AI Property Image Platform
# ============================================================================

# ----------------------------------------------------------------------------
# Deep Learning (PyTorch ecosystem)
# Install with: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# Or for CPU: pip install torch torchvision torchaudio
# ----------------------------------------------------------------------------
# torch>=2.0.0
# torchvision>=0.15.0
# torchaudio>=2.0.0

# ----------------------------------------------------------------------------
# HuggingFace Ecosystem
# ----------------------------------------------------------------------------
transformers>=4.40.0
datasets>=2.19.0
accelerate>=0.29.0
timm>=0.9.16
huggingface-hub>=0.22.0

# ----------------------------------------------------------------------------
# Computer Vision
# ----------------------------------------------------------------------------
opencv-python>=4.9.0
Pillow>=10.3.0
albumentations>=1.4.0
scikit-image>=0.23.0

# ----------------------------------------------------------------------------
# Image Enhancement
# ----------------------------------------------------------------------------
basicsr>=1.4.2
realesrgan>=0.3.0
facexlib>=0.3.0
gfpgan>=1.3.8

# ----------------------------------------------------------------------------
# AI Agents and LLM
# ----------------------------------------------------------------------------
langchain>=0.1.20
langchain-core>=0.1.52
langchain-community>=0.0.38
langgraph>=0.0.55
google-genai==0.8.0
openai>=1.25.0

# ----------------------------------------------------------------------------
# Vector Database
# ----------------------------------------------------------------------------
qdrant-client>=1.9.0

# ----------------------------------------------------------------------------
# Backend API
# ----------------------------------------------------------------------------
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.2.0

# ----------------------------------------------------------------------------
# Database and Storage
# ----------------------------------------------------------------------------
supabase>=2.4.0
sqlalchemy>=2.0.32
alembic>=1.13.0
psycopg2-binary>=2.9.9
redis>=5.0.0

# ----------------------------------------------------------------------------
# ML Utilities
# ----------------------------------------------------------------------------
scikit-learn>=1.4.0
numpy>=1.26.0,<2.0.0
pandas>=2.2.0
wandb>=0.17.0

# ----------------------------------------------------------------------------
# Visualization
# ----------------------------------------------------------------------------
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.21.0

# ----------------------------------------------------------------------------
# Notebook
# ----------------------------------------------------------------------------
jupyter>=1.0.0
ipykernel>=6.29.0
ipywidgets>=8.1.0

# ----------------------------------------------------------------------------
# Utilities
# ----------------------------------------------------------------------------
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.66.0
rich>=13.7.0
click>=8.1.0
requests>=2.31.0
aiofiles>=23.2.0
httpx>=0.27.0
schedule>=1.2.0
